fix: avoid Terraform and CodeDeploy fighting (#26)
## Context
Fixes #25 

The blue/green deployment of the ingest task fails when redeployed due
to an issue in the ALB's target group configuration of the corresponding
listener rule, which is what CodeDeploy uses to switch from an existing
task set to a newly deployed one.

## Proposed Changes
Add a lifecycle meta-argument to the rule to avoid Terraform undoing
configuration done by CodeDeploy, and making CodeDeploy think that
something is broken.

## Tests
I did a brand new deployment and then added some changes and redeployed
and the ingest task deployment went through flawlessly the second time
too.